extract Edit Handler
