+++ gitit2.cabal
@@ -55,6 +55,7 @@ library
                    Network.Gitit2.Handler.Category
                    Network.Gitit2.Handler.Delete
                    Network.Gitit2.Handler.Diff
+                   Network.Gitit2.Handler.Edit
                    Network.Gitit2.Handler.History
                    Network.Gitit2.Handler.Random
                    Network.Gitit2.Handler.Search
